Adding Final Comments
This code worked well for ~50 computers on the floor, was able to make the desired changes in about 2 passes. I'm thinking this saved me about a day or two of work, otherwise I would have to physically go to the computer and click through the options to make the changes above. This also saved employees time since the changes could be made without them stopping work.